In the search of El-Farol bar problem
I analysed the outcome of a famous dance competition show using the
record of dancers' performances, recorded as scores awarded by a judging
panel and the public's sentiment, reflected in tweets on Twitter.
My study first shows that the outcome of the voting mechanisms with
publicly known signals can be understood using a game theoretic model
known as the El-Farol game.
And the second, it shows that the relationship between the wisdom of
crowds and the outcomes of a voting structure cannot be described by a
linear function and can therefore lead to very contradictory results.